// src/app/modules/authorization/types/authorization.types.ts
export enum Action {
  Manage = 'manage',
  Create = 'create',
  Read = 'read',
  Update = 'update',
  Delete = 'delete',
  ViewAnalytics = 'viewAnalytics',
  ManageCustomDomain = 'manageCustomDomain',
  ManageWhiteLabel = 'manageWhiteLabel',
}

export enum AuthorizationSubject {
  LinkPage = 'LinkPage',
  LinkPageAnalytics = 'LinkPageAnalytics',
  CompanyDomain = 'CompanyDomain',
  LinkPageWhiteLabel = 'LinkPageWhiteLabel',
}